Show git commit ID before running the TS.
Dockerfile updated to set a new envar inside the container image
pointing to the TNF repo commit id/sha/tag that was checked out
to be included in the image. This id/tag will be displayed
by the run-cnf-suites.sh before launching the test suites.

In case the script run-cnf-suites.sh is launched locally, outside of the
tnf container, it will show "unknown".