add default timeout to send command in spawned shell (#356)
add default timeout to send command in spawned shell

This aims to add timeout to "Send" function in goexpect library, by default, the library does not use any timeout.
By looking in the code of the goexpect, it's printing the log after "sending" the command, so if there's any issue in the "send" we will not see the log. This seems like the issue we were seeing this morning where the process hangs.