CRD autodiscovery using spec.group.
Added targetCrds yaml entry for CRDs autodiscovery. Currently, they can
only be discovered using the group (spec.group) tag, but it can be
extended to use others like "name" or just labels as with pods/cuts.

It just creates a list with the names of the CRDs that belong to the
configured groups.