Insert claim.json file in results HTML page
To simplify log viewing, we want to have the CI job insert the
claim.json file automatically inside the results.html page. This cannot
be done by dynamically loading it due to CORS restrictions, but we can
convert the JSON file into a javascript file, then load it from the HTML
page.