Lifecycle's TC for terminationGracePeriod param refactored. (#562)
* Lifecycle's TC for terminationGracePeriod param refactored.

Iterates through deployments/statefulsets under tests to get the
last-applied-configuration field value. If the param was set in the yaml
file, it should appear there. Otherwise, the default value will be used
by k8s, but the TC fails.

* Added support for unmanaged pods.

CNFs can also have unmanaged/orphan pods, which are those that do not
belong to any deployment/statefulset. This is not a recommened way to
deploy pods, but it's allowed (see the lifecycle's Owners TC, which
is just informative).

* Fixed final results processing.

Co-authored-by: Jun Chen <jun@redhat.com>