#1 enhance catalog browsing (#561)
* just a first attempt to get suites and names

* structure and filling a new catalog

* new structure created for better printing
 basically the structure allows to iterate
 and print the catalogue divided by
 # suite-test
 ## test-name

* base domain contant for tests urls

to have a connstant as a base domain for all the
test cases bulding blocks.

This can be used later to extract the test name
from an url composed with the base domain

* slight changes for the new formatting

* printing Test Cases Building Blocks

  these are slightly different from Test Cases

* fix version not printing on Test Cases Building

* fixed bullet points indentation

* Erros managment for identifiers url

check that url identifiers are ok based on the
configured base domain for every set of tests

* the new catalog style

* fixing linting issues

* functionality to get test name from Identier.URL

* functionality to get suite and test names from identifier

* test for the new function XformToGinkgoItIdentifier

* new format for the catalog

* append directly a catalogElement object

Clear code in that way

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>

* cleaner structs assignment

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>

* cleaner assignment

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>

* cleaner assigments of elements into slices

* forgotten debug statement

* cleaner assignment

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>

* cleaner assigment

* cleaner return of suiteTest

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>

* solved import issue

* improved some functinons documentatoin

* test case building block no need array of identifiers

* wrong and ok test combined in the same slice

* updated version

* fixed lint issure

* changes in printing atrributes

* no need this line

* Just adding some clarifications

* interchanged actualy by expected in assert

* fixed markdown new line

* simplified outputTestCaseBuildingBlocks

actually is mostly as it was at the beggining. I
complicated this function very when it was no need it.

I had only to extract the testName from the URL dont need
to make it so complex, as it is in outputTestCases

* better name for getting name from identifier

"test building blocks have nothing to do with ginkgo "  Jun is right.
now I see the differences between Test cases and Test cases buidling
blocks. So is makes sense to call the function to get name from
identifier in a more proper way. Nothing related to Ginkgo nor pointing
to the need of skiping tests. Because here you dont skip tests

Co-authored-by: Brandon Palm <brandonpalm@gmail.com>