Allow diagnostic mode when no ginkgo focus were set.
This commit adds a new flag for the binary called "diagnostic"
(-diagnostic) so in case it's found, only diagnostic and
autodiscovery-related information will be dumped into the claim file.
No tests specs will be run when the flag is set. The launching script
will automatically set the -diagnostic flag in case the focus
flag (-f) was not used.